% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseCohort_functions.R
\name{FitRakingModel_CC}
\alias{FitRakingModel_CC}
\title{Case-cohort versions of raking estimators}
\usage{
FitRakingModel_CC(valid_dat, dat_sim, mod)
}
\arguments{
\item{valid_dat}{Validation subset dataset}

\item{dat_sim}{Full dataset}

\item{mod}{String indicating which model to get influence
functions from}
}
\value{
Raking model fit object
}
\description{
These functions implement raking estimators
when the validation subset was selected using
a case-cohort sampling scheme.
}
