% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{CalcCorrectedlogHR}
\alias{CalcCorrectedlogHR}
\title{Runs RRCME procedure}
\usage{
CalcCorrectedlogHR(full_dat, valid_dat, num_boot = 250, sampling_type,
  estimators = c("RC", "RSRC", "GRRC", "GRN"))
}
\arguments{
\item{full_dat}{Data for full cohort 
containing error-prone variables}

\item{valid_dat}{Validation subset
from full cohort containing
error-free variables}

\item{num_boot}{Number of bootstrap 
replicates}

\item{sampling_type}{Sampling design
for validation subset}

\item{estimators}{Vector of strings 
indicating which estimators to run}
}
\value{
Matrix of estimates and 
standard errors
}
\description{
Runs some combination of Regression
Calibration, Risk set Regression
Calibration, Generalized Raking
Regression Calibration, and 
Generalized Raking Naive as specified 
by the estimators input. 
Log hazard ratios and corresponding
standard errors are returned.
}
