% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking_estimators.R
\name{FitRakingModel}
\alias{FitRakingModel}
\title{Calculates raking estimates}
\usage{
FitRakingModel(valid_dat, full_dat, mod, sampling_type)
}
\arguments{
\item{valid_dat}{Validation subset dataset}

\item{full_dat}{Full dataset}

\item{mod}{String indicating which model to get influence
functions from}

\item{sampling_type}{String indicating either simple random
sampling or case-cohort sampling}
}
\value{
Raking model fit object
}
\description{
These functions implement the raking estimators. First, either
a naive model or RC model is fit. Then, influence functions from
that model is extracted and then used as auxiliary variables
to improve efficiency of Horvitz-Thompson estimates.
}
