% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_estimators.R
\name{FitRCModel}
\alias{FitRCModel}
\alias{CalcExpX}
\alias{CalcExpw}
\title{Calculates regression calibration estimates}
\usage{
FitRCModel(valid_dat, full_dat, sampling_type, return_coef)

CalcExpX(valid_dat, full_dat, sampling_type)

CalcExpw(valid_dat, full_dat, sampling_type)
}
\arguments{
\item{valid_dat}{Validation subset}

\item{full_dat}{Full dataset}

\item{sampling_type}{String indicating either simple random
sampling or case-cohort sampling}

\item{cox_time}{Censored time-to-event}

\item{cox_delta}{Failure indicator}

\item{cox_x}{Univariate covariate measured
with error}

\item{cox_z}{Univariate covariate measured
without error}
}
\value{
Eiher a list of Cox model coefficients
and standard errors of a Cox model fit object
}
\description{
These functions implement the regression
calibration estimators. Moment estimators
are fit using least squares and are used
to obtain our best prediction of the
true covariate and censored event time.
}
