% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSRC_estimators.R
\name{FitRSRCModel}
\alias{FitRSRCModel}
\title{Calculates regression calibration estimates}
\usage{
FitRSRCModel(valid_dat, dat_sim, sampling_type, beta_x_start, beta_z_start)
}
\arguments{
\item{valid_dat}{Validation subset}

\item{dat_sim}{Full dataset}

\item{sampling_type}{String indicating either simple random
sampling or case-cohort sampling}

\item{beta_x_start}{Initial guess for beta_x in optimization}

\item{beta_x_start}{Initial guess for beta_z in optimization}
}
\value{
List of RSRC beta_x and beta_z estimates
}
\description{
These functions implement the risk set
regression calibration estimators.
Moment estimators are fit using least
squares and are used to obtain our best
prediction of the true covariate and
censored event time. This is repeated
at deciles of the failure times.
}
